Refactor TUI session into input and queue modules
